requirements.txt
addict
scipy
tensorboard
scikit-learn
lmdb
pillow
tensorboard
jupyter
jupyterlab
matplotlib
opencv-python
torchvision=0.4.0
